pyproject.toml
[tool.poetry]
name = "raia"
version = "0.1.0"
description = "discord bot"
authors = ["JohnCarloCarl19 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
python-dotenv = "^0.18.0"
Flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
